added functionality for cutting planes in texture based vis. Once switch node is implemented in our graph we'll be good to go for scalar data. There are still issues about matching the names up w/ the classic side and getting the correct enum. Right now texture based vis uses *_contours to switch cutting planes. . .

+++ VE_Xplorer/cfdVolumeVisualization.h
@@ -46,7 +46,7 @@ public:
    void SetTextureManager(cfdTextureManager* tm);
    void UpdateStateSet(osg::StateSet* ss);
    void CreateNode();
-   void AddClipPlane(ClipPlane direction,float* position);
+   void AddClipPlane(ClipPlane direction,double* position);
    void RemoveClipPlane(ClipPlane direction);
    void UpdateClipPlanePosition(ClipPlane direction,double* newPosition);
    bool isCreated(){return _isCreated;}
